[tests] update benchmark tests, config
The benchmarks were showing slow results for AsyncFileHandler,much
slower compared to what they were until v0.4.0. The reason is that
a couple of bugs were recently fixed in that handler, that resulted
in the following configurable property, used for the benchmark tests,
to not have the desired effect:

org.eclipse.tracecompass.traceeventlogger.AsyncFileHandler.maxSize = 1000

After the bugs were fixed, AsyncFileHandler started to correctly use
"maxSize" as the size of its record buffer. However, a buffer of 1000
is not enough for the logging load generated by the benchmark tests,
making the new async handler look bad.

Setting "bufferSize" to value 100k makes it big enough to "absorb" the
logging load of the benchmark tests well, and brings back the previous
level of performance for the async handler, such that the benchmark
results are comparable to those from until v0.4.0.

To detect such a problem in the future, the benchmark tests now compute
the relative performance of the async file handler vs the old sync
handler and fails the test if the former is not perfoeming above a
certain threshold. The threshold may need to be lowered a bit if e.g.
it's not met once in a while during CI (there is variation in the
results)

In the benchmark results printed on STDOUT, the relative performance of
the async file handleris now displayed as an extra column in the
benchmark result tables.

Also, the benchmark results are now printed in two versions: one un-
formated in "csv" format that's suitable to import e.g. in LibreOffice
Calc, and another that's formatted for human readability.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>